split wordle into command-view-model-stats
